refactor: move discovery fn to its own file
